+++ litebird_sim/simulations.py
@@ -1548,7 +1548,7 @@ class Simulation:
     @_profile
     def add_noise(
         self,
-        random: np.random.Generator,
+        random: Union[np.random.Generator, None] = None,
         noise_type: str = "one_over_f",
         component: str = "tod",
         append_to_report: bool = True,
@@ -1558,10 +1558,13 @@ class Simulation:
         This method must be called after having set the instrument,
         the list of detectors to simulate through calls to
         :meth:`.set_instrument` and :meth:`.add_detector`.
-        The parameter `random` must be specified and must be a random number
-        generator thatimplements the ``normal`` method. You should typically
-        use the `random` field of a :class:`.Simulation` object for this.
+        The parameter `random` can be specified as a random number
+        generator that implements the ``normal`` method. As default it uses
+        the `random` field of a :class:`.Simulation` object for this.
         """
+
+        if random is None:
+            random = self.random
 
         add_noise_to_observations(
             observations=self.observations,